moved post request handler to controller
