+++ routes/upload.js
@@ -4,7 +4,6 @@ const router = express.Router();
 const UploadController = require("../controllers/upload");
 
 router.get("/", UploadController.Index);
-// router.post("/", UploadController.UploadImage);
-// router.get("/new", ProfileController.New);
+router.post("/", UploadController.UploadImage);
 
 module.exports = router;
